pyproject.toml
[project]
name = "AbstractImgGen"
version = "0.1.0"
description = "Abstract Image generator using hill climbing optimizations."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.1.3",
    "pillow>=11.0.0",
]
